// src/shared/enums/ForecastVariableDaily.ts
export enum ForecastVariableDaily {
  APPARENT_TEMPERATURE_MAX = "apparent_temperature_max",
  APPARENT_TEMPERATURE_MIN = "apparent_temperature_min",
  ET0_FAO_EVAPOTRANSPIRATION = "et0_fao_evapotranspiration",
  PRECIPITATION_HOURS = "precipitation_hours",
  PRECIPITATION_SUM = "precipitation_sum",
  RAIN_SUM = "rain_sum",
  SHORTWAVE_RADIATION_SUM = "shortwave_radiation_sum",
  SHOWERS_SUM = "showers_sum",
  SNOWFALL_SUM = "snowfall_sum",
  SUNRISE = "sunrise",
  SUNSET = "sunset",
  TEMPERATURE_2M_MAX = "temperature_2m_max",
  TEMPERATURE_2M_MIN = "temperature_2m_min",
  WEATHERCODE = "weathercode",
  WINDDIRECTION_10M_DOMINANT = "winddirection_10m_dominant",
  WINDGUSTS_10M_MAX = "windgusts_10m_max",
  WINDSPEED_10M_MAX = "windspeed_10m_max",
}